drop redundant high-order handler
